requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
requests==2.31.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
openai==1.14.3
pydantic==2.6.4
httpx==0.24.1
psycopg2-binary==2.9.11
